refactor(react-components): program-list のバリアントを index.tsx に統合
ProgramList / ProgramListLoading / ProgramListError を 1 ファイル
(index.tsx) に併記する形に統合。stories も index.stories.tsx 1 本に集約
(通常表示・ローディング・エラー・フローの全パターン)。
barrel index.ts と旧 variant 別ファイル・各 stories を削除。

internal な program-list-item.tsx / program-list-grid.tsx /
create-program-card.tsx は別ファイルのまま変更なし。
