--- packages/react-components/src/views/program-list/index.ts
@@ -1,3 +1,0 @@
-export { ProgramList } from "./program-list";
-export { ProgramListError } from "./program-list-error";
-export { ProgramListLoading } from "./program-list-loading";
