--- packages/react-components/src/views/program-list/program-list-loading.tsx
@@ -1,28 +1,0 @@
-import type { FC } from "react";
-import { CreateProgramCard } from "./create-program-card";
-import { ProgramListGrid, ProgramListGridItem } from "./program-list-grid";
-
-type Props = {
-	createHref: string;
-};
-
-const SKELETON_KEYS = ["s1", "s2", "s3", "s4", "s5"] as const;
-
-export const ProgramListLoading: FC<Props> = ({ createHref }) => {
-	return (
-		<ProgramListGrid aria-busy aria-label="プログラムを読み込み中">
-			<ProgramListGridItem>
-				<CreateProgramCard href={createHref} />
-			</ProgramListGridItem>
-			{SKELETON_KEYS.map((key) => (
-				<ProgramListGridItem
-					key={key}
-					className="h-full min-h-20 rounded-lg bg-overlay p-4 shadow-sm"
-				>
-					<div className="h-5 w-2/3 animate-pulse rounded bg-muted" />
-					<div className="mt-2 h-3 w-1/3 animate-pulse rounded bg-muted" />
-				</ProgramListGridItem>
-			))}
-		</ProgramListGrid>
-	);
-};
